+++ R/extract_dsd.R
@@ -14,7 +14,9 @@
 #'    options(restatapi_cores=2)
 #' }    
 #' }
-#' dsd_url<-"https://ec.europa.eu/eurostat/SDMX/diss-web/rest/datastructure/ESTAT/DSD_nama_10_a10_e"
+#' \donttest{
+#' dsd_url<-"https://ec.europa.eu/eurostat/SDMX/diss-web/rest/datastructure/ESTAT/DSD_med_rd6"
+#' options(timeout=2)
 #' tryCatch({
 #'   dsd_xml<-xml2::read_xml(dsd_url)}, 
 #'   error=function(e){
@@ -22,7 +24,8 @@
 #'   warning=function(w){
 #'   message("Unable to download the xml file.\n",w)}) 
 #' if (exists("dsd_xml")) {extract_dsd("GEO",dsd_xml)} 
-#' 
+#' options(timeout=2)
+#' }
 
 extract_dsd<-function(concept=NULL,dsd_xml=NULL){
   if (is.null(dsd_xml)|is.null(concept)){