+++ R/get_eurostat_data.R
@@ -121,6 +121,7 @@
 #' }    
 #' }
 #' \donttest{
+#' options(timeout=2)
 #' dt<-get_eurostat_data("NAMA_10_GDP")
 #' dt<-get_eurostat_data("htec_cis3",update_cache=TRUE,check_toc=TRUE)
 #' dt<-get_eurostat_data("agr_r_milkpr",cache_dir="/tmp",cflags=TRUE)
@@ -163,6 +164,7 @@
 #'                        label=TRUE,
 #'                        name=FALSE)
 #' clean_restatapi_cache("/tmp",verbose=TRUE)                                 
+#' options(timeout=60)
 #' }
 
 get_eurostat_data <- function(id,
@@ -189,7 +191,9 @@ get_eurostat_data <- function(id,
   update_cache<-update_cache|getOption("restatapi_update",FALSE)
   dmethod<-getOption("restatapi_dmethod",get("dmethod",envir=restatapi::.restatapi_env))
   tbc<-cr<-TRUE # to be continued for the next steps  / cache result data.table 
-  # options(code_opt=NULL)
+  if (verbose) {message("\nget_eurostat_data - footer code option value at start:",paste(getOption("code_opt",NULL),collapse=", "))}
+  options(code_opt=NULL)
+  if (verbose) {message("get_eurostat_data - footer code option value after reset:",paste(getOption("code_opt",NULL),collapse=", "))}
   if(cflags){keep_flags<-cflags}
  
   if (!(exists(".restatapi_env"))) {load_cfg()}
@@ -372,7 +376,9 @@ get_eurostat_data <- function(id,
             if (tbc & !is.null(xml_foot)){
                 code<-xml2::xml_attr(xml_foot,"code")
                 if (length(code)!=0){
-                  message("The query had at least one footer message. You can check the details with the verbose=TRUE parameter.")
+                  notification<-"The query had at least one footer message."
+                  if (!verbose) {notification<-paste(notification,"You can check the details with the 'verbose=TRUE' parameter.")}
+                  message(notification)
                 }
                 severity<-xml2::xml_attr(xml_foot,"severity")
                 fmsg<-xml2::xml_text(xml2::xml_children(xml_foot))
@@ -403,15 +409,16 @@ get_eurostat_data <- function(id,
             if (file.exists(temp)) unlink(temp)
             if (!is.null(rdat)){data.table::as.data.table(rdat,stringsAsFactors=stringsAsFactors)}
         }),fill=TRUE)
+        if (verbose) {message("get_eurostat_data - footer code option value after retrieval:",paste(getOption("code_opt",NULL),collapse=", "))}
         if (!is.null(restat)) #at least one url provided a valid result => check if all the queries with data downloaded
         {
           if ((nrow(restat)==0)) #  if there is no data in the results 
           { 
-            if (all(getOption("code_opt",NULL)==500)) # no data for all the filters => stop processing by restat<-NULL
+            if (all(getOption("code_opt",NULL)==404)) # no data for all the filters => stop processing by restat<-NULL
             { 
-              message("500 - No data with the given filter(s)")
+              message("404 - No data retrived with the given filter(s)")
               restat<-NULL
-            } else #there is some data but was not downloaded
+            } else if (any(getOption("code_opt",NULL)==413))  #there is some data but was not downloaded
             {   
               if (local_filter) #apply filter locally
               { 
@@ -433,7 +440,13 @@ get_eurostat_data <- function(id,
               { 
                 message("No data retrieved for the given filter(s), because the results are too big to download immediately through the REST API. You may want to download the whole dataset and apply the filter(s) locally.")
               }
-            }
+            } else if (any(getOption("code_opt",NULL)>=500)) #if there is some data but for some of the filters there is a warning that there is "internal application error" or "exception while getting all data and footnotes slice" or "Cannot connect to Comext service."
+              {
+                notification<-"One or some of the filter(s) resulted a warning or error footer message in the response of the REST API. The retrieved partial data is discarded."
+                if (!verbose) {notification<-paste(notification,"You can check the details rerunning the request with the 'verbose=TRUE' parameter.")}
+                message(notification)
+                restat<-NULL
+              }  
           } else if (any(getOption("code_opt",NULL)==413)) #if there is some data but for some of the filters there is a warning that could not be downloaded imediately
           { 
             if (local_filter) #apply filter locally and replace the data from the REST API
@@ -452,7 +465,13 @@ get_eurostat_data <- function(id,
               if (nrow(restat)>0){message("The retrived dataset is partial!!!")}
               message("One or some of the filter(s) resulted too large datatset to download through the REST API. You may want to download the whole dataset and apply the filter(s) locally.")
             }
-          } 
+          } else if (any(getOption("code_opt",NULL)>=500)) #if there is some data but for some of the filters there is a warning that there is "internal application error" or "exception while getting all data and footnotes slice" or "Cannot connect to Comext service."
+          {
+            notification<-"One or some of the filter(s) resulted a warning or error footer message in the response of the REST API. The retrieved partial data is discarded."
+            if (!verbose) {notification<-paste(notification,"You can check the details rerunning the request with the 'verbose=TRUE' parameter.")}
+            message(notification)
+            restat<-NULL
+          }  
           cr<-FALSE # do not cahce filtered data only bulk datasets
         }
       }